Refactor/dev

See merge request react-node-track/class01/web-project1/team04/help_me_backend!20

+++ src/controller/email.controller.js
@@ -7,6 +7,8 @@ export const emailController = {
       const userEmail = req.user.email;
       const email = await emailService.getReceivedEmail(userEmail);
 
+      if (!userEmail) throw new Error ('Bad Request+유저 정보를 찾을 수 없음');
+
       res.status(200).json(email);
     } catch(e) {
       next(e);
@@ -19,6 +21,8 @@ export const emailController = {
       const userEmail = req.user.email;
       const email = await emailService.getSentEmail(userEmail);
 
+      if (!userEmail) throw new Error ('Bad Request+유저 정보를 찾을 수 없음');
+
       res.status(200).json(email);
     } catch(e) {
       next(e);
@@ -30,6 +34,8 @@ export const emailController = {
     try {
       const userEmail = req.user.email;
       const newEmailCount = await emailService.checkNewEmail(userEmail);
+
+      if (!userEmail) throw new Error ('Bad Request+유저 정보를 찾을 수 없음');
 
       res.status(200).json({
         newEmailCount, // 새로운 이메일 개수 반환
@@ -47,6 +53,8 @@ export const emailController = {
       const userEmail = req.user.email;
       const email = await emailService.checkedEmail(email_id, userEmail);
 
+      if (!userEmail) throw new Error ('Bad Request+유저 정보를 찾을 수 없음');
+
       res.status(200).json(email);
     } catch(e) {
       next(e);
@@ -59,6 +67,8 @@ export const emailController = {
     try {
       const { title, content, target_email } = req.body;
       const userEmail = req.user.email;
+
+      if (!userEmail) throw new Error ('Bad Request+유저 정보를 찾을 수 없음');
 
       if (!title || !content) throw new Error('Bad Request+제목이나 내용이 없습니다.');
 
@@ -78,6 +88,8 @@ export const emailController = {
     try {
       const { email_id } = req.params;
       const userEmail = req.user.email;
+
+      if (!userEmail) throw new Error ('Bad Request+유저 정보를 찾을 수 없음');
 
       const result = await emailService.deleteEmail(email_id, userEmail);
